requirements.txt
paramiko==3.4.0
websockets==12.0
SQLAlchemy==2.0.27
python-dotenv==1.0.1
rich==13.7.0
fastapi==0.110.0
uvicorn==0.27.1
aiofiles==23.2.1
jinja2==3.1.3 